LayoutTests/http/tests/media/hls/hls-audio-tracks-has-audio.html

<!DOCTYPE html>
<html>
    <head>
        <script src=../../media-resources/video-test.js></script>
        <script src=../../media-resources/media-controls.js></script>
        <script>
            function start()
            {
                findMediaElement();
                waitForEvent('canplaythrough', canplaythrough);
                video.src = "../resources/hls/audio-tracks.m3u8";
            }

            function canplaythrough()
            {
                testExpected("window.internals.mediaElementHasCharacteristic(video, 'Audible')", true);
                endTest();
            }
        </script>
    </head>
    <body onload="start()">
        <video id="video"></video>
    </body>
</html>
