Safari-611.1.10/LayoutTests/http/tests/media/hls/hls-audio-tracks-has-audio-expected.txt


EVENT(canplaythrough)
EXPECTED (window.internals.mediaElementHasCharacteristic(video, 'Audible') == 'true') OK
END OF TEST
